chore: enforce format, lint, and typecheck CI quality gates
Add Prettier and ESLint workspace tooling, gate Python/TypeScript formatting,
linting, and typechecking in ci-cd.yml, and fix remaining violations so
make format-check, lint, and typecheck pass across the monorepo.

+++ apps/backend/src/config/version_formatting.py
@@ -124,4 +124,4 @@ def format_elevation(elevation_m: float, version: str) -> int:
         Formatted elevation value
     """
     rounding = get_elevation_rounding(version)
-    return round(elevation_m, rounding)
+    return int(round(elevation_m, rounding))
